feat(hummock): unified test in hummock iterator
* skeleton

* implement TestIteratorBuilder

* implement test iterator

* implement test_validator

* rewrite sorted test

* use u64 table_id

* use assert_hummock_value helper

* rewrite test in user_key_test

* fix clippy && fmt errors

* remove redundant comment

* fix rebase errors

* stronger test

+++ rust/storage/src/hummock/iterator/concat.rs
@@ -122,7 +122,7 @@ mod tests {
         iter.rewind().await.unwrap();
 
         while iter.is_valid() {
-            let table_idx = i / TEST_KEYS_COUNT;
+            let table_idx = (i / TEST_KEYS_COUNT) as u64;
             let key = iter.key();
             let val = iter.value();
             assert_eq!(
